+++ src/edu/jhuapl/sbmt/core/rendering/PerspectiveImageBackplanesHelper.java
@@ -249,7 +249,7 @@ class PerspectiveImageBackplanesHelper
                 double[] x = new double[3];
                 double[] pcoords = new double[3];
                 int[] subId = new int[1];
-                int[] cellId = new int[1];
+                long[] cellId = new long[1];
                 int result = cellLocator.IntersectWithLine(spacecraftPositionAdjusted[currentSlice], lookPt, tol, t, x, pcoords, subId, cellId, cell);
 
                 // if (intersectPoints.GetNumberOfPoints() == 0)
@@ -268,7 +268,7 @@ class PerspectiveImageBackplanesHelper
                     // double[] closestPoint = intersectPoints.GetPoint(0);
                     // int closestCell = intersectCells.GetId(0);
                     double[] closestPoint = x;
-                    int closestCell = cellId[0];
+                    int closestCell = (int)cellId[0];
                     double closestDist = MathUtil.distanceBetween(closestPoint, spacecraftPositionAdjusted[currentSlice]);
 
                     /*
